+++ ParticleToy/forces/WallForce.h
@@ -5,12 +5,12 @@
 
 class WallForce : public Force{
 public:
-    WallForce(Particle *p1, double fixed_x);
+    WallForce(Particle *p1, float fixed_x);
     void calc_Force() override;
     void draw() override;
 
 private:
 
     Particle *const m_p1;   // particle 1
-    double const m_x;
+    float const m_x;
 };
